fitting/requirement.txt
smplx==0.1.28
Flask==3.1.0
numpy==1.26.4
roma==1.5.1
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
opencv-python==4.10.0.84
pyav==11.4.1
tqdm==4.67.1
statsmodels==0.14.4
xformers==0.0.27
matplotlib==3.10.0
gdown==5.2.0
huggingface-hub==0.27.0
#triton==2.3.1
pyrender==0.1.45
einops==0.8.0
lapx==0.5.11.post1
cython_bbox==0.1.5
tensorboard==2.18.0
pytorch-lightning==2.2.4
timm==1.0.3
dill==0.3.8
webdataset==0.2.100
omegaconf==2.3.0
git+https://github.com/IXRPlateform/chumpy
roma==1.5.1
yacs==0.1.8
scenedetect[opencv]==0.6.5.2
rich==13.9.4
ultralytics==8.3.65
gql==3.6.0b3
librosa==0.10.2.post1
fastapi==0.115.8
websockets==14.2
easydict==1.13